README.md
As of late 2021 Russia has blocked the use of Tor by many of their major ISP's and as such a lot of blogs will now be offline as many RAAS gangs are russian.

RAAS Group Blogs To Stay up to Date with the latest leaks. The goal is to help companies quickly identify where leaks are posted.